[@mantine/lightbox] Lightbox: Fix ESM import throwing on a circular dependency between Lightbox and LightboxProvider
